Downgrade license from AGPLv3 to regular GPLv3
There are only 8 commits in cabal-helper not by me or alanz, see:

    $ git log --author='Daniel Gröber\|Alan Zimmerman' --invert-grep -p
    9142d8a9e6ed18faf17a360521fbbbd25f6a3b47

Alan Zimmerman has agreed to the license change on IRC.

The rest are IMO uncopyrightable. I'm pretty sure version-bounds are "facts"
which are inherently uncopyrightable, i.e. the software builds with this and
that version, that's a fact. Simple build fixes such as ae568d10 and eb178a5b
shouldn't really cross the necessary creative threshold.